Added code for user

+++ donHub/donHub/src/main/java/com/donHub/donHub/repository/UserRepositoryI.java
@@ -2,8 +2,8 @@ package com.donHub.donHub.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import com.donHub.donHub.model.ProductRequest;
+import com.donHub.donHub.model.UserRequest;
 
-public interface UserRepositoryI extends MongoRepository<UserRequest, Integer>{
+public interface UserRepositoryI extends MongoRepository<UserRequest, Long>{
 
 }
